added job for archiver
