+++ config/application.rb
@@ -11,5 +11,6 @@ module Guild
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
+    config.active_job.queue_adapter = :sidekiq
   end
 end
