+++ spec/models/job_spec.rb
@@ -7,12 +7,16 @@ RSpec.describe Job, :type => :model do
       expect(newclass).to be_a(described_class)
     end
   end
+  describe "#initialization" do
+    let(:job){ Job.new }
+    it "should set the default archived boolean to false" do
+      expect( job.archived ).to eq false
+    end
+  end
+
   describe "#archive" do
     let(:job){ Job.new }
     it "should set the default archived boolean to false" do
-
-
-
       expect( job.archived ).to eq false
     end
   end
